+++ src/components/modal/z-login-modal/index.tsx
@@ -12,6 +12,7 @@ import {
   InputStatusBean,
   InputStatusEnum,
 } from "../../../beans";
+import { getSiblings } from "../../../utils/utils";
 
 /**
  * @slot username - username input
@@ -39,35 +40,10 @@ export class ZLoginModal {
   @State() externalProviderCheck: boolean = false;
 
   componentDidLoad() {
-    console.log(
-      this.hostElement.querySelector('slot-fb[name="provider"]').children.length
-    );
-    this.externalProviderCheck = !!this.getSiblings(
-          this.hostElement.querySelector('slot-fb[name="provider"]')
-        ).filter((sibl) => sibl.nodeName !== 'Z-BODY').length
-
-        // console.log('check :>> ', check);
-      }
-
-
-  private getSiblings(elem: HTMLElement) {
-    // for collecting siblings
-    let siblings = [];
-    // if no parent, return no sibling
-    if (!elem.parentNode) {
-      return siblings;
-    }
-    // first child of the parent node
-    let sibling = elem.parentNode.firstChild;
-
-    // collecting siblings
-    while (sibling) {
-      if (sibling.nodeType === 1 && sibling !== elem) {
-        siblings.push(sibling);
-      }
-      sibling = sibling.nextSibling;
-    }
-    return siblings;
+    this.externalProviderCheck = !!getSiblings(
+      // INFO: lo shadowDOM è aperto, si possono usare i selettori standard
+      this.hostElement.querySelector('slot-fb[name="provider"]')
+    ).filter((sibl) => sibl.nodeName !== "Z-BODY").length;
   }
 
   /** Emitted on login submit */
@@ -253,21 +229,23 @@ export class ZLoginModal {
             <z-body class="provider" level={5} variant="regular">
               OPPURE ACCEDI CON:
             </z-body>
-            {!this.externalProviderCheck && <slot name={`provider`}>
-              <z-button variant={ButtonVariantEnum.secondary}>
-                PROVIDER
-              </z-button>
-              <z-button variant={ButtonVariantEnum.secondary}>
-                PROVIDER
-              </z-button>
-              {this.renderZainoDigitaleButton()}
-              <z-link
-                icon="informationsource"
-                href="https://www.zainodigitale.it/#/landing"
-              >
-                Cos'è Zaino Digitale?
-              </z-link>
-            </slot>}
+            {!this.externalProviderCheck && (
+              <slot name={`provider`}>
+                <z-button variant={ButtonVariantEnum.secondary}>
+                  PROVIDER
+                </z-button>
+                <z-button variant={ButtonVariantEnum.secondary}>
+                  PROVIDER
+                </z-button>
+                {this.renderZainoDigitaleButton()}
+                <z-link
+                  icon="informationsource"
+                  href="https://www.zainodigitale.it/#/landing"
+                >
+                  Cos'è Zaino Digitale?
+                </z-link>
+              </slot>
+            )}
           </div>
         </div>
       </z-modal>
