+++ src/utils/utils.ts
@@ -59,3 +59,20 @@ export function getElementTree(
 
   return tree;
 }
+
+export function getSiblings(elem: HTMLElement) {
+  // for collecting siblings
+  let siblings = [];
+  // if no parent, return no sibling
+  if (!elem.parentNode) return siblings;
+  // first child of the parent node
+  let sibling = elem.parentNode.firstChild;
+  // collecting siblings
+  while (sibling) {
+    if (sibling.nodeType === 1 && sibling !== elem) {
+      siblings.push(sibling);
+    }
+    sibling = sibling.nextSibling;
+  }
+  return siblings;
+}
